+++ crates/sandogasa-bugzilla/src/lib.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+pub mod claim;
 mod client;
 pub mod models;